Add files via upload
This is a script to run in Adobe Illustrator. It saves a Illustrator 8 copy next to your .ai File that can be opened in C4D. The Copy will be named "c4d_export" + exportCount + ".ai". If the source .ai File has not been saved somewhere yet the export File will be saved to your desktop.